Extract UI rendering
