--- universal-login-web3/lib/ui.tsx
@@ -1,5 +1,0 @@
-import React from 'react'
-
-const App = () => <div>Hello, world!</div>
-
-export default <App />
