+++ universal-login-web3/lib/ui/App.tsx
@@ -1,0 +1,4 @@
+import React from 'react';
+
+export const App = () => <div>Hello, world!</div>;
+
